Add Test for Vector, Matrix and LU classes
`test_lu.cc` script contains a main() method.
The method uses the 3 classes implemented for Vector, Matrix and LU, and all methods requested as part of the exercise, to solve a Matrix-Vector-Equation for an unknown vector x.
By default, the method prints out the result for x, as well as a test for a residual $ where  = b - Ax$.